C++ script for image processing an pose initialization.
This C++ script undistorts the 3 camera images, tests various image binarization techniques, and generates an initial object pose relative to camera 3 using the centre of the bounding rectangle drawn around the convex hull of the tub perimeter in the image. It also projects a 3D model of the tub into the image but the optimised pose is not computed in this script.